requirements.txt
google-generativeai==0.7.2
python-telegram-bot==21.5